requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
aiomysql==0.2.0
python-dotenv==1.0.0
google-genai==0.3.0
httpx==0.25.2
greenlet==2.0.2
PyMySQL==1.1.0
